Finish modules refactoring
